+++ src/client/configs/parts/GridEditorModal.tsx
@@ -2,43 +2,38 @@ import { memo, useEffect } from 'react';
 import GridData from '@logic-pad/core/data/grid';
 import { cn } from '../../../client/uiHelper.ts';
 import EmbedContext, { useEmbed } from '../../contexts/EmbedContext.tsx';
-import PuzzleEditor from '../../editor/PuzzleEditor';
-import GridContext, {
-  GridConsumer,
-  useGrid,
-} from '../../contexts/GridContext.tsx';
+import PuzzleEditorScreen from '../../screens/PuzzleEditorScreen.tsx';
+import GridContext, { GridConsumer } from '../../contexts/GridContext.tsx';
 import DisplayContext from '../../contexts/DisplayContext.tsx';
 import EditContext from '../../contexts/EditContext.tsx';
 import GridStateContext from '../../contexts/GridStateContext.tsx';
 
-export interface GridEdittorModalProps {
+export interface GridEditorModalProps {
   grid: GridData;
   setGrid: (grid: GridData) => void;
   open: boolean;
   setOpen: (open: boolean) => void;
 }
 
-function EmbedLoader({ grid }: { grid: GridData }) {
+function EmbedLoader() {
   const { setFeatures } = useEmbed();
-  const { setGrid } = useGrid();
   useEffect(() => {
     setFeatures({
       instructions: false,
       metadata: false,
       checklist: false,
     });
-    setGrid(grid, null);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
   return null;
 }
 
-export default memo(function GridConfig({
+export default memo(function GridEditorModal({
   grid,
   setGrid,
   open,
   setOpen,
-}: GridEdittorModalProps) {
+}: GridEditorModalProps) {
   return (
     <dialog id="gridModal" className={cn('modal', open && 'modal-open')}>
       <div className="modal-box w-[calc(100%-4rem)] h-full max-w-none bg-neutral text-neutral-content">
@@ -58,9 +53,9 @@ export default memo(function GridConfig({
                 <DisplayContext>
                   <EditContext>
                     <GridStateContext>
-                      <GridContext>
-                        <EmbedLoader grid={grid} />
-                        <PuzzleEditor>
+                      <GridContext grid={grid} solution={null}>
+                        <EmbedLoader />
+                        <PuzzleEditorScreen>
                           <GridConsumer>
                             {({ grid, setGrid: setInnerGrid }) => {
                               return (
@@ -88,7 +83,7 @@ export default memo(function GridConfig({
                               );
                             }}
                           </GridConsumer>
-                        </PuzzleEditor>
+                        </PuzzleEditorScreen>
                       </GridContext>
                     </GridStateContext>
                   </EditContext>