+++ src/client/contexts/GridContext.tsx
@@ -41,15 +41,25 @@ export const GridConsumer = context.Consumer;
 
 export default memo(function GridContext({
   children,
+  grid: initialGrid,
+  solution: initialSolution,
+  metadata: initialMetadata,
 }: {
   children: React.ReactNode;
+  grid?: GridData | (() => GridData);
+  solution?: GridData | null | (() => GridData | null);
+  metadata?: PuzzleMetadata;
 }) {
   const { recordEdit, clearHistory } = useEdit();
   const { setState } = useGridState();
 
-  const [grid, setGrid] = useState(defaultGrid);
-  const [solution, setSolution] = useState<GridData | null>(null);
-  const [metadata, setMetadata] = useState<PuzzleMetadata>(defaultMetadata);
+  const [grid, setGrid] = useState(initialGrid ?? defaultGrid);
+  const [solution, setSolution] = useState<GridData | null>(
+    initialSolution === undefined ? null : initialSolution
+  );
+  const [metadata, setMetadata] = useState<PuzzleMetadata>(
+    initialMetadata ?? defaultMetadata
+  );
 
   useEffect(() => {
     clearHistory(grid);
@@ -60,13 +70,21 @@ export default memo(function GridContext({
     newGrid.symbols.forEach(list => {
       list.forEach(symbol => {
         if (handlesSetGrid(symbol)) {
-          newGrid = symbol.onSetGrid(grid, newGrid);
+          newGrid = symbol.onSetGrid(
+            grid,
+            newGrid,
+            sol === undefined ? solution : sol
+          );
         }
       });
     });
     newGrid.rules.forEach(rule => {
       if (handlesSetGrid(rule)) {
-        newGrid = rule.onSetGrid(grid, newGrid);
+        newGrid = rule.onSetGrid(
+          grid,
+          newGrid,
+          sol === undefined ? solution : sol
+        );
       }
     });
     setGrid(newGrid);