+++ src/client/routes/_context._layout.perfection.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from '@tanstack/react-router';
+import { validateSearch } from '../router/linkLoader';
+
+export const Route = createFileRoute('/_context/_layout/perfection')({
+  validateSearch,
+});